+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG
 
+## [0.3.0](https://github.com/zCloak-Network/zkid-login/compare/v0.2.0...v0.3.0) (2022-12-11)
+
+
+### Bug Fixes
+
+* **providers:** zkid wallet not install ([4520ffe](https://github.com/zCloak-Network/zkid-login/commit/4520ffe28543a9f5ed124527fca0bc902ac26984))
+
+
+### Features
+
+* **zkid-provider:** add generate zkp ([#9](https://github.com/zCloak-Network/zkid-login/issues/9)) ([ce0df05](https://github.com/zCloak-Network/zkid-login/commit/ce0df05102ed951e6b37090fc04642c38703b568))
+
+
 ## [0.2.0](https://github.com/zCloak-Network/zkid-login/compare/v0.1.1...v0.2.0) (2022-12-09)